refactor: delete `AuthenticationConfiguration` and test, update refs to use CSL equivalents
