--- security/security-core/src/main/java/io/camunda/security/configuration/AuthenticationConfiguration.java
@@ -1,72 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.security.configuration;
-
-import io.camunda.security.api.model.config.AuthenticationMethod;
-import io.camunda.security.api.model.config.oidc.OidcConfiguration;
-import io.camunda.security.api.model.config.oidc.OidcProvidersConfiguration;
-
-public class AuthenticationConfiguration {
-  public static final AuthenticationMethod DEFAULT_METHOD = AuthenticationMethod.BASIC;
-  public static final boolean DEFAULT_UNPROTECTED_API = false;
-
-  private AuthenticationMethod method = DEFAULT_METHOD;
-  private String authenticationRefreshInterval = "PT30S";
-  private OidcConfiguration oidcAuthenticationConfiguration = new OidcConfiguration();
-  private boolean unprotectedApi = DEFAULT_UNPROTECTED_API;
-
-  private OidcProvidersConfiguration providers;
-
-  public boolean getUnprotectedApi() {
-    return unprotectedApi;
-  }
-
-  public void setUnprotectedApi(final boolean value) {
-    unprotectedApi = value;
-  }
-
-  public AuthenticationMethod getMethod() {
-    return method;
-  }
-
-  public void setMethod(final AuthenticationMethod method) {
-    this.method = method;
-  }
-
-  public String getAuthenticationRefreshInterval() {
-    return authenticationRefreshInterval;
-  }
-
-  public void setAuthenticationRefreshInterval(final String authenticationRefreshInterval) {
-    this.authenticationRefreshInterval = authenticationRefreshInterval;
-  }
-
-  public OidcConfiguration getOidc() {
-    return oidcAuthenticationConfiguration;
-  }
-
-  public void setOidc(final OidcConfiguration configuration) {
-    oidcAuthenticationConfiguration = configuration;
-  }
-
-  public OidcProvidersConfiguration getProviders() {
-    return providers;
-  }
-
-  public void setProviders(final OidcProvidersConfiguration providers) {
-    this.providers = providers;
-  }
-
-  public boolean isCamundaGroupsEnabled() {
-    return !(getMethod() == AuthenticationMethod.OIDC && getOidc().isGroupsClaimConfigured());
-  }
-
-  public boolean isCamundaUsersEnabled() {
-    return getMethod() != AuthenticationMethod.OIDC;
-  }
-}
