+++ security/security-core/src/main/java/io/camunda/security/configuration/SecurityConfigurations.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.security.configuration;
 
+import io.camunda.security.api.model.config.AuthenticationConfiguration;
 import io.camunda.security.api.model.config.AuthenticationMethod;
 
 public class SecurityConfigurations {
